+++ src/time/mod.rs
@@ -138,3 +138,6 @@ pub use uptime::*;
 
 mod htop;
 pub use htop::*;
+
+mod time_unit;
+pub use time_unit::*;